# Fix Array type instability