Adresses #9.

I changed occurrences of `::Array{Something}` to `::Vector{Something}` in internal data structures' definitions.
Someone may want to check that I did not put a `Vector` where there should have been a `Matrix`; all unit tests pass locally though.

There is still `units::Array` and `price_sensitive_loads::Array` in the definition of `Bus`. I tried to changed these to `Vector{Unit}` and `Vector{PriceSensitiveLoad}`, but it errors at compilation because those types haven't been defined yet.
I don't believe forward declaration of types is possible in Julia at the moment.

___

Here's an example illustrating the effects of the proposed modification.
The function below simply extracts the minimum power cost for each generator and each time period. On the current `dev`, this allocates ~170kB of memory. On the revised branch, the same code allocates 0kB and runs a 100x faster.

```julia
using BenchmarkTools
using LinearAlgebra
BLAS.set_num_threads(1)

using UnitCommitment

instance = UnitCommitment.read_benchmark("matpower/case1354pegase/2017-02-01")

function get_min_power_cost!(Z, uc::UnitCommitmentInstance)
    T = uc.time
    G = length(uc.units)
    for (g, gen) in enumerate(uc.units)
        for t in 1:T
            Z[g, t] = gen.min_power_cost[t]
        end
    end
    return Z
end

T = instance.time
G = length(instance.units)
Z = zeros(G, T)

println("UC instance has $G units and $T time periods")

println("Extracting min_power_cost matrix")
b = @benchmark get_min_power_cost!($Z, $instance)
display(b)
println()
```

Before:
```
UC instance has 260 units and 36 time periods
Extracting min_power_cost matrix
BenchmarkTools.Trial: 
  memory estimate:  174.70 KiB
  allocs estimate:  10141
  --------------
  minimum time:     796.400 μs (0.00% GC)
  median time:      889.650 μs (0.00% GC)
  mean time:        920.316 μs (0.96% GC)
  maximum time:     3.437 ms (73.81% GC)
  --------------
  samples:          5426
  evals/sample:     1
```
With the changes:
```
UC instance has 260 units and 36 time periods
Extracting min_power_cost matrix
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     8.967 μs (0.00% GC)
  median time:      9.167 μs (0.00% GC)
  mean time:        9.313 μs (0.00% GC)
  maximum time:     60.567 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     3
```